[Bronze V] Title: Multiply, Time: 172 ms, Memory: 17660 KB, Score: 40 point -BaekjoonHub

+++ Java/백준/Bronze/22193. Multiply/Multiply.java
@@ -4,13 +4,13 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int n = scanner.nextInt();
-        int m = scanner.nextInt();
+        long n = scanner.nextInt();
+        long m = scanner.nextInt();
 
-        int x = scanner.nextInt();
-        int y = scanner.nextInt();
+        long x = scanner.nextInt();
+        long y = scanner.nextInt();
 
-        int sum = x * y;
+        long sum = x * y;
 
         System.out.println(sum);
